move storage files
