+++ lib/storage/__mocks__/index.js
@@ -1,6 +1,6 @@
-import MemoryOnlyProvider, {mockStore, mockSet, setMockStore} from '../../storage/providers/MemoryOnlyProvider';
+import MemoryOnlyProvider, {mockStore, mockSet, setMockStore} from '../providers/MemoryOnlyProvider';
 
-const MultiStorageMock = {
+const StorageMock = {
     init: jest.fn(MemoryOnlyProvider.init),
     getItem: jest.fn(MemoryOnlyProvider.getItem),
     multiGet: jest.fn(MemoryOnlyProvider.multiGet),
@@ -19,4 +19,4 @@ const MultiStorageMock = {
     setMockStore: jest.fn((data) => setMockStore(data)),
 };
 
-export default MultiStorageMock;
+export default StorageMock;
