+++ lib/storage/index.js
@@ -1,6 +1,7 @@
 import _ from 'underscore';
-import PlatformStorage from '../storage';
-import MemoryOnlyProvider from '../storage/providers/MemoryOnlyProvider';
+import PlatformStorage from './platforms';
+import MemoryOnlyProvider from './providers/MemoryOnlyProvider';
+import keepInstancesSync from './keepInstancesSync/index.web';
 
 let storageProviders = [MemoryOnlyProvider, PlatformStorage];
 
@@ -45,6 +46,8 @@ const MultiStorage = {
     },
 
     init() {
+        this.keepInstancesSync = keepInstancesSync.bind(this);
+
         // If a storage provider fails to initialize, we remove it from the list of providers
         // We don't want to remove the "memory only" provider though, therefore, we start at index 1
         return Promise.all(
