+++ lib/storage/keepInstancesSync/index.native.js
@@ -1,0 +1,6 @@
+/**
+ * Noop on native
+ */
+function keepInstancesSync() {}
+
+export default keepInstancesSync;
