+++ lib/storage/platforms/NativeStorage.js
@@ -1,7 +1,7 @@
-import SQLiteStorage from './providers/SQLiteStorage';
+import SQLiteProvider from '../providers/SQLiteProvider';
 
 const webStorage = {
-    ...SQLiteStorage,
+    ...SQLiteProvider,
 
     /**
      * Noop on native
