+++ lib/storage/platforms/WebStorage.js
@@ -1,0 +1,7 @@
+import IDBKeyValProvider from '../providers/IDBKeyValProvider';
+
+const webStorage = {
+    IDBKeyValProvider,
+};
+
+export default webStorage;
