+++ lib/storage/platforms/index.native.js
