+++ lib/storage/providers/SQLiteProvider.js
